feat(map): implement curved paths and SNR-weighted lines for traceroutes
